Rewrite `addEdge` to use binary search to insert the element in the already sorted slice.

Benchmark for `Insert`:

```
name       old time/op    new time/op    delta
Insert-10     266ns ± 1%     233ns ± 2%  -12.33%  (p=0.029 n=4+4)

name       old alloc/op   new alloc/op   delta
Insert-10      161B ± 0%      137B ± 0%  -14.91%  (p=0.029 n=4+4)

name       old allocs/op  new allocs/op  delta
Insert-10      4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.029 n=4+4)
```